# docs/swagger.yaml
definitions:
  controllers.CategoryInput:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  controllers.SubcategoryInput:
    properties:
      category:
        type: string
      name:
        type: string
    type: object
...
  models.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Subcategory:
    properties:
      category_id:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.SubcategoryResponse:
    properties:
      category:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: List of categories
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
      summary: Get all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Creates a new category with the provided name
      parameters:
      - description: Category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/controllers.CategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: Category created
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Validation Error
          schema:
            type: string
      summary: Create a new category
      tags:
      - categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an existing category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category successfully deleted
        "404":
          description: Category not found
          schema:
            type: string
      summary: Delete a category
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Retrieves a category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category found
          schema:
            $ref: '#/definitions/models.Category'
        "404":
          description: Category not found
          schema:
            type: string
      summary: Get a category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Updates the name of an existing category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/controllers.CategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Validation Error
          schema:
            type: string
        "404":
          description: Category not found
          schema:
            type: string
      summary: Update a category
      tags:
      - categories
  /subcategories:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all subcategories with their categories
      produces:
      - application/json
      responses:
        "200":
          description: List of subcategories
          schema:
            items:
              $ref: '#/definitions/models.SubcategoryResponse'
            type: array
      summary: Get all subcategories
      tags:
      - subcategories
    post:
      consumes:
      - application/json
      description: Creates a new subcategory within a category
      parameters:
      - description: Subcategory creation data
        in: body
        name: subcategory
        required: true
        schema:
          $ref: '#/definitions/controllers.SubcategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: Subcategory created
          schema:
            $ref: '#/definitions/models.Subcategory'
        "400":
          description: Invalid JSON payload or validation error
          schema:
            type: string
        "403":
          description: Unknown category
          schema:
            type: string
      summary: Create a new subcategory
      tags:
      - subcategories
  /subcategories/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an existing subcategory by ID
      parameters:
      - description: Subcategory ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Subcategory successfully deleted
        "404":
          description: Subcategory not found
          schema:
            type: string
      summary: Delete a subcategory
      tags:
      - subcategories
    get:
      consumes:
      - application/json
      description: Retrieves a subcategory by its ID including category name
      parameters:
      - description: Subcategory ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Subcategory found
          schema:
            $ref: '#/definitions/models.SubcategoryResponse'
        "404":
          description: Subcategory not found
          schema:
            type: string
      summary: Get a subcategory by ID
      tags:
      - subcategories
    put:
      consumes:
      - application/json
      description: Updates an existing subcategory by ID
      parameters:
      - description: Subcategory ID
        in: path
        name: id
        required: true
        type: integer
      - description: Subcategory update data
        in: body
        name: subcategory
        required: true
        schema:
          $ref: '#/definitions/controllers.SubcategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: Subcategory updated
          schema:
            $ref: '#/definitions/models.Subcategory'
        "400":
          description: Invalid JSON payload or validation error
          schema:
            type: string
        "403":
          description: Unknown category
          schema:
            type: string
        "404":
          description: Subcategory not found
          schema:
            type: string
      summary: Update a subcategory
      tags:
      - subcategories
...
swagger: "2.0"